% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add.ellipsis}
\alias{add.ellipsis}
\title{Add Ellipsis}
\usage{
add.ellipsis(f)
}
\description{
Adds the ... argument to a function. Useful when ... is forwarded down to functions that don't take the ellipsis as argument
}
