% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{add.rank.column}
\alias{add.rank.column}
\title{Adding ranking columns}
\usage{
add.rank.column(T, name.cols, prefix = "rnk", pct = FALSE, by = NULL)
}
\arguments{
\item{T}{a data.table}

\item{name.cols}{the columns to be ranked}

\item{prefix}{the prefix that will be appended to the ranking columns}

\item{pct}{Should the ranking be expressed in percentage}

\item{by}{A grouping to do the ranking in}
}
\description{
Adds a rank column for the column specified
}

