% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{adjugate}
\alias{adjugate}
\alias{minor}
\alias{cofactor}
\alias{adjugate1}
\alias{adjugate2}
\title{Minor, Cofactor, Adjugates}
\usage{
minor(A, i, j)

cofactor(A, i, j)

adjugate1(A)

adjugate2(A)
}
\arguments{
\item{A}{Matrix}

\item{i}{reference line}

\item{j}{reference column}
}
\description{
Tools to compute the minor, cofactor or adjugate matrix of a given matrix
}
\details{
From stackoverflow Vincent Zoonekynd: http://stackoverflow.com/questions/16757100/get-adjoint-matrix-in-r
}
\section{Functions}{
\itemize{
\item \code{minor}: Minor, determinant where row i and j are taken out

\item \code{cofactor}: Cofactor, minor multiplied by (-1) ^ (i + j)

\item \code{adjugate1}: Adjugate Matrix. Adjugate(M)[i, j] = Cofactor(M, j, i)

\item \code{adjugate2}: Adjugate Matrix computed without the use of for loops but with outer and Vectorize
}}

\references{
https://en.wikipedia.org/wiki/Adjugate_matrix
}
