% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{align.factors}
\alias{align.factors}
\title{Aligning factor levels}
\usage{
align.factors(..., facts = list(...))
}
\arguments{
\item{...}{Any number of factor vectors}

\item{facts}{a list of factor vectors}
}
\description{
Transform a list of factors into a list of factors with compatible levels
}
