% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{array.index}
\alias{array.index}
\title{Array Index}
\usage{
array.index(..., LI = list(...), dim, one.based = TRUE)
}
\arguments{
\item{...}{integer vectors describing the coordinates}

\item{LI}{a list of the individual coordinates}

\item{dim}{Dimension of the array}

\item{one.based}{should the indices start at one}
}
\description{
Computes the index in the vectorization of a multidimensional array corresponding to the individual coordinates
}
