% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{array.to.data.table}
\alias{array.to.data.table}
\title{Array to data.table}
\usage{
array.to.data.table(A, one.based = TRUE)
}
\arguments{
\item{A}{Array to transform to a data.table}

\item{one.based}{Should the indices start at one}
}
\description{
Transforms an array into a set of index / value tuples
}
