% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.vectorization.R
\name{assign.array.vect}
\alias{assign.array.vect}
\title{Assigning data by index value pair}
\usage{
assign.array.vect(M, I, values, dimn = dim(M))
}
\description{
Assigning data by index value pair
}
