% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge_sparse.R
\name{autocrossprod}
\alias{autocrossprod}
\title{Auto crossprod}
\usage{
autocrossprod(x, ...)
}
\arguments{
\item{x}{Matrix}

\item{...}{Additional settings}
}
\description{
Generic function that will implement \code{crossprod} and \code{tcrossprod} in the case x = y
}

