% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.vectorization.R
\name{bdiag.to.vech}
\alias{bdiag.to.vech}
\title{Vectorization of block diagonal matrix}
\usage{
bdiag.to.vech(...)
}
\arguments{
\item{vdim}{vector of size of each block}

\item{logger}{JLogger to be used to log messages}
}
\description{
Creates the matrix that transforms the concatenation of each blocks half vectorization into the a half vectorization of the entire matrix
}
\seealso{
diag.to.vech, diag.to.vec, bdiag.to.vec
}
