% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{bdiag.with.names}
\alias{bdiag.with.names}
\title{Matrix rows and column names}
\usage{
bdiag.with.names(L)
}
\description{
Build a block diagonal matrix and computes the row names and column names from the block names
}
