% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{cartesian.data.table}
\alias{cartesian.data.table}
\title{Cartesian product of two data.tables}
\usage{
cartesian.data.table(D1, D2 = D1, size.limit = Inf, ...)
}
\arguments{
\item{D1}{data.table}

\item{D2}{data.table}

\item{size.limit}{A limit on the size of the result. If the result exceeds that size data is randomly sampled from D1 and D2}

\item{...}{Not used}
}
\description{
Will create a table of size nrow(D1) * nrow(D2) consisting of all the possible combinations of rows in D1 and D2
}
