% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{cartesian.list}
\alias{cartesian.list}
\title{Cartesian product of a several data.tables}
\usage{
cartesian.list(L, ...)
}
\arguments{
\item{L}{list of data.tables}

\item{...}{Parameters to be forwarded to cartesian.data.table}
}
\description{
Cartesian product of a several data.tables
}
