% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{col.bind}
\alias{col.bind}
\title{Resize, reserve, binds}
\usage{
col.bind(M1, M2)
}
\description{
Binds two data structures so that the number of columns of the result is the sum of the number of columns of the two arguments. Number of rows has to be te same
}
\details{
Some objects are meant to be growing as some algorithms run. These functions provides feature to handle these cases
}

