% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{col.classes}
\alias{col.classes}
\title{Column classes}
\usage{
col.classes(D, cols = names(D))
}
\arguments{
\item{D}{table}
}
\description{
Shortcut to get the class of each columns of the table
}

