% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{compare.matrices}
\alias{compare.matrices}
\title{Comparing Matrices}
\usage{
compare.matrices(M1, M2)
}
\arguments{
\item{M1}{A matrix}

\item{M2}{A matrix}
}
\value{
A list with tree elements:
\itemize{
\item{n1} L2 Norm of M1
\item{n2} L2 Norm of M2
\item{n12} Scalar product of M1 and M2 normalized by M1's norm
}
}
\description{
Compares two matrices by computing their norms and computing there scalar products normalized bby M1's norm.
}
