% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.conversion.R
\name{convert.generic.date}
\alias{convert.generic.date}
\title{Converts Date}
\usage{
convert.generic.date(date, date.class, origin = "1970-01-01")
}
\arguments{
\item{date}{date vector}

\item{date.class}{class to be converted to}

\item{origin}{origin for type conversion fonction}
}
\description{
Converts various types of Date to a tractable one
}
