% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.save.objects.R
\name{copy.mc.cmd}
\alias{copy.mc.cmd}
\title{Copy file from/to s3 storage}
\usage{
copy.mc.cmd(
  src,
  dest,
  ...,
  retrieved.output = FALSE,
  recursive = FALSE,
  quiet = FALSE,
  logger = ukpik.logger()
)
}
\arguments{
\item{src}{full path of the source file}

\item{dest}{full path of the saving destination}

\item{...}{}

\item{retrieved.output}{if mc output should be retrieved to R}

\item{recursive}{should be TRUE of src is a directory}

\item{quiet}{set to TRUE for removing logs}

\item{logger}{JLogger}
}
\description{
Copy file from/to s3 storage
}
\section{Functions}{
\itemize{
\item \code{copy.mc.cmd()}: copy file or directory from s3 storage to local storage or the opposite

}}
