% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{cor.dec}
\alias{cor.dec}
\title{Covariance matrix to correlation matrix}
\usage{
cor.dec(S, tol = 10 * .Machine$double.eps)
}
\description{
Computes the correlation matrix associated with a given covariance matrix
}
\seealso{
cov2cor
}

