% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{create.cartesian.dt}
\alias{create.cartesian.dt}
\title{Cartesian product of several vectors}
\usage{
create.cartesian.dt(...)
}
\arguments{
\item{...}{vectors that will be used to compute the cartesian product}
}
\description{
Cartesian product of several vectors
}
