% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.sendmail.R
\name{critical.mail}
\alias{critical.mail}
\title{Critical mail}
\usage{
critical.mail(from = guess.user(), to, subject, msg, ...,
  logger = NULL)
}
\arguments{
\item{from}{Email from}

\item{to}{Recipient}

\item{subject}{Subject of the email}

\item{msg}{Message}

\item{logger}{JLogger to be used to log messages}
}
\description{
Tool to send email from R
}
