% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{deep.list}
\alias{deep.list}
\title{List assignment}
\usage{
deep.list(L = list(), keys, value)
}
\arguments{
\item{L}{the list}

\item{keys}{Vector of the keys. A path in the tree formed in the}

\item{value}{Optional value to be assigned}
}
\description{
Request or assigns values deep in the list at any arbitrary depth.
}
\examples{
L = deep.list(list(), c("A", "B", "C"), 1)
deep.list(L, c("A", "B", "C"))
}
