% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{df.partial.kronecker}
\alias{df.partial.kronecker}
\title{Partial Kronecker Product}
\usage{
df.partial.kronecker(DM1, DM2, Il = NULL, Ir = NULL, dm2)
}
\arguments{
\item{DM1}{a triplet representation of a matrix}

\item{DM2}{DITTO}

\item{Il}{The list of row indices in the result that we are interested in keeping. If NULL assumes no projection. Il must be replicated locally to each process.}

\item{Ir}{The list of column indices in the result that we are interested in keeping. If NULL assumes no projection. Ir must be replicated locally to each process.}

\item{dm2}{Dimension of the RHS matrix}
}
\description{
Computes a kronecker product for a limited subset of the rows and columns. This version returns the triplet representation as a \code{data.frame}
}
\details{
Kronecker products produce are very costly in terms of memory usage.
}
