% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.vectorization.R
\name{diag.to.vec}
\alias{diag.to.vec}
\title{Vectorization of a diagonal matrix}
\usage{
diag.to.vec(n, logger = jlogger::JLoggerFactory("jalgos filter"))
}
\arguments{
\item{n}{Size of the diagonal matrix}

\item{logger}{JLogger to be used to log messages}
}
\description{
Creates the matrix that transforms the vector of diagonal value into the vectorization of the whole matrix
}
\seealso{
diag.to.vech, bdiag.to.vec, bdiag.to.vech
}
