% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.vectorization.R
\name{diag.to.vech}
\alias{diag.to.vech}
\title{Vectorization of a diagonal matrix}
\usage{
diag.to.vech(n, logger = jlogger::JLoggerFactory("jalgos filter"))
}
\arguments{
\item{n}{Size of the diagonal matrix}

\item{logger}{JLogger to be used to log messages}
}
\description{
Creates the matrix that transforms the vector of diagonal value into the half vectorization of the whole matrix
}
\seealso{
diag.to.vec, bdiag.to.vec, bdiag.to.vec
}

