% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.beautfication.R
\name{din}
\alias{din}
\title{Default Value if NULL provided}
\usage{
din(val, default)
}
\arguments{
\item{val}{Value to test for nullity}

\item{default}{Value to return if null}
}
\description{
Returns `default` if `val` is null
}
