% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{dsparseMatrix}
\alias{dsparseMatrix}
\title{Building Matrices}
\usage{
dsparseMatrix(D, ...)
}
\arguments{
\item{D}{triplet representation, needs to have i, j, x columns}

\item{...}{Parameters to be forwarded to sparseMatrix}
}
\value{
a Matrix of class Matrix
}
\description{
Builds a sparseMatrix from a data.table containing a triplet representation of a matrix. The function will only use the i, j, x columns
}
