% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{dt.cart.apply}
\alias{dt.cart.apply}
\title{Cartesian Apply}
\usage{
dt.cart.apply(funapp, Largs, ...)
}
\arguments{
\item{funapp}{Function to apply}

\item{Largs}{list of parameters}
}
\description{
Applies a function for all possible combinations of a set of parameters using a data.table
}
