% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{dt.grep}
\alias{dt.grep}
\alias{dtgrep}
\alias{grep.merge}
\alias{amerge}
\title{Merging using grep}
\usage{
dtgrep(D, vals, akeys, fgrep = grep, keep.match.col = FALSE, ...)

grep.merge(D1, D2, keys = character(0), akeys = character(0),
  fgrep = grep, ...)

amerge(D1, D2, keys = character(0), akeys = character(0), fgrep = grep,
  similarity.fun, ...)
}
\arguments{
\item{D}{A data.table}

\item{vals}{Values to match. A list with one entry per column}

\item{akeys}{Approximate keys. List of columns to match on.}

\item{fgrep}{Grep function to use}

\item{keep.match.col}{Should the columns used to match be included in the result}

\item{D1}{A data.table}

\item{D2}{A data.table}

\item{keys}{Keys to exactly match on}
}
\description{
These are tools to merge data.tables using grep instead of an exact match
}
\section{Functions}{
\itemize{
\item \code{dtgrep}: Returns a subset of D which matches vals for the columns akeys

\item \code{grep.merge}: Merges with grep

\item \code{amerge}: Approximate merge
}}

