% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{DT.matching.cols}
\alias{DT.matching.cols}
\title{Data.table matching columns}
\usage{
DT.matching.cols(DT, match.exp, ig.match.exp = "^$", ...,
  logger = emergency.logger())
}
\arguments{
\item{DT}{a data table}

\item{match.exp}{the regular expression to be matched by column names}

\item{ig.match.exp}{the regular expression to be unmatched by column names}
}
\description{
Retrieve the data.table with columns that match a pattern AND does not match an other pattern
}
