% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_geocoding.R
\name{earth.dist}
\alias{earth.dist}
\title{Earth Distance}
\usage{
earth.dist(lon1, lat1, lon2, lat2)
}
\arguments{
\item{lon1}{Longitude of the first point}

\item{lat1}{Latitude of the second point}

\item{lon2}{Longitude of the first point}

\item{lat2}{Latitude of the second point}
}
\value{
Second order approximation of the distance between the two points in kilometers.
}
\description{
Second order estimation of the distance between two points parametrized by latitude and longitude
}

