% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{eigen.func}
\alias{eigen.func}
\alias{funcM}
\alias{sqrtM}
\alias{absM}
\alias{posD}
\alias{posM}
\alias{expM}
\alias{logM}
\title{Applying functions to the eigen value}
\usage{
funcM(M, func, ...)

sqrtM(M, eps = .Machine$double.eps)

absM(M)

posD(D, ...)

posM(M, ...)

expM(M, ...)

logM(M, ...)
}
\arguments{
\item{M}{Matrix on which we want to apply func}

\item{func}{Function to apply to the eigen values of M}

\item{...}{Parameters to pass on to func}

\item{D}{eigen values}
}
\section{Functions}{
\itemize{
\item \code{funcM}: Applies a function to the eigen decomposition of a matrix in this fashion: let M = Pdiag(V)P^-1 then the result is Q = Pdiag(f(V))P^-1 where f is a function that is applied to the vector of eigen values.

\item \code{sqrtM}: The function applied is the square root. A tolerance can be given to consider negative values close to 0 to be 0

\item \code{absM}: Absolute value of the eigen values

\item \code{posD}: Positive value

\item \code{posM}: Positive value of the eigen values. In case of a symmetric matrix, it will turn it into a semi-definite positive matrix.

\item \code{expM}: Exponential of the eigen values. Computes the exponential of a matrix. It is however not the most efficient manner.

\item \code{logM}: Computes the logarithm of a matrix. Best defined in the case of definite positive matrices
}}

