% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{empty.column}
\alias{empty.column}
\title{Building a data.table}
\usage{
empty.column(type, N)
}
\arguments{
\item{type}{string that defines the type}

\item{N}{length of the column}
}
\description{
Creates empty columns of a given type
}
