% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{eqfactors}
\alias{eqfactors}
\title{Test equality between factors}
\usage{
eqfactors(F1, F2)
}
\arguments{
\item{F1}{factor or character vector}

\item{F2}{factor or character vector}
}
\description{
Tests equality for factor vectors with possibly different levels. This is simply done by converting the factors to character.
}

