% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{evalLand}
\alias{evalLand}
\title{Evaluating valid subsets}
\usage{
evalLand(Tab, Filts)
}
\arguments{
\item{Tab}{a data.table}

\item{Filts}{list of expressions or logical vectors.}
}
\description{
Evaluate several filters related to a 'data.table'
}
