% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.tests.R
\name{expect_equal_matrices}
\alias{expect_equal_matrices}
\title{Matrix Equality Expectation}
\usage{
expect_equal_matrices(M1, M2, compare.dimnames = TRUE)
}
\arguments{
\item{compare.dimnames}{Should dimnames be compared as well}
}
\description{
Tests whether two matrix represent the same data by converting them to a triplet representation.
}
\details{
Calls to \code{expect_equal} yield to unexpected results. If the goal of the comparison is only to check whether the conceptual matrices represented by M1 and M2 are identical this does the trick.
}

