% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{filter.na}
\alias{dt.filter.na}
\alias{dt.filter.out.na}
\alias{filter.na}
\title{Filtering non-finite values}
\usage{
dt.filter.na(DT, numeric.cols = names(DT), good.fun = is.finite, ...)

dt.filter.out.na(DT, ...)
}
\arguments{
\item{DT}{Table to be filtered}

\item{numeric.cols}{Numeric columns in the table, defaults to the entire table}

\item{good.fun}{Function that filters out bad values. It should return FALSE for the unwanted values. It will be applied on each column one by one}

\item{...}{Extra parameters to be passed on to 'dt.filter.na' and 'good.fun'}
}
\value{
'dt.filter.na' returns the set of bad rows while 'dt.filter.out.na' returns a data.table where the bad rows have been filtered out
}
\description{
Computed the set of rows that have a non-finite value in the numeric columns.
}

