% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{first.last}
\alias{first}
\alias{first.last}
\alias{last}
\title{Easy access to first and last element}
\usage{
first(X)

last(X)
}
\arguments{
\item{X}{a vector or a list}
}
\section{Functions}{
\itemize{
\item \code{first}: first element

\item \code{last}: last element
}}

