% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{flip}
\alias{flip}
\title{Flipping a data.table}
\usage{
flip(DT, by)
}
\arguments{
\item{DT}{a data.table}

\item{by}{Columns to be ommited}
}
\description{
Transposes a data.table so the columns become the rows and vice versa
}
