% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{funcM}
\alias{funcM}
\title{Factors Handling}
\usage{
funcM(M, ...)
}
\arguments{
\item{M}{Matrix to be decomposed}

\item{func}{Function to be applied to the eigen values}
}
\description{
Applies a function to the eigen decomposition of a matrix in this fashion: let M = Pdiag(V)P^-1 then the result is Q = Pdiag(f(V))P^-1 where f is a function that is applied to the vector of eigen values.
}
