% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fusion.factors}
\alias{fusion.factors}
\title{Binding factor vectors}
\usage{
fusion.factors(F1, F2, lev = union(levels(F1), levels(F2)))
}
\arguments{
\item{F1}{factor vector}

\item{F2}{factor vector}

\item{lev}{New levels}
}
\description{
Binds two factor vectors. Will take care of setting the new levels. Binding with a call to \code{c} returns a integer vector
}
