% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.non.finite.R
\name{gen.is.finite}
\alias{gen.is.finite}
\title{Check Finiteness For All Types}
\usage{
gen.is.finite(x)
}
\arguments{
\item{x}{vector to check}
}
\description{
Returns true for each entry if x is finite. Can take factors and characters as well
}
