% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{get.dt.names}
\alias{get.dt.names}
\title{Get column names using regex}
\usage{
get.dt.names(reg.names, DT, ignore.case = TRUE, value = TRUE, ...)
}
\arguments{
\item{reg.names}{regex to use}

\item{DT}{data.table}

\item{ignore.case}{sets case tolerance, default is case doesn't matter}

\item{value}{returning indices or names, defaults to names}

\item{...}{forwarded to grep}
}
\description{
Will get the column names of a table that match a given regex
}
