% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\docType{methods}
\name{ginv}
\alias{ginv}
\alias{ginv,Matrix-method}
\title{Generalized inverse}
\usage{
ginv(X, tol = sqrt(.Machine$double.eps))

\S4method{ginv}{Matrix}(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{Matrix to invert}

\item{tol}{espsilon bump to force full rank}
}
\description{
Computes the generalized inverse for any type of matrix: rectangular and not full rank.
}

