% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\docType{methods}
\name{gsolve}
\alias{gsolve}
\alias{gsolve,Matrix-method}
\title{Generalized inverse}
\usage{
gsolve(a, b, ...)

\S4method{gsolve}{Matrix}(a, b, eps = sqrt(.Machine$double.eps))
}
\arguments{
\item{a}{Matrix to invert}

\item{b}{RHS of the equation a.v = b where v is the unknown}
}
\description{
Computes a solution of the a.v = b for any type of matrix: rectangular and not full rank.
}

