% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.hash.R
\name{hash.combine}
\alias{hash.combine}
\title{Combining Hash Numbers}
\usage{
hash.combine(h1, h2)
}
\arguments{
\item{h1}{Vector of hash numbers}

\item{h2}{Vector of hash numbers}
}
\description{
Combines two hash numbers into one
}

