% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge_sparse.R
\docType{class}
\name{HugeMatrix-class}
\alias{HugeMatrix-class}
\title{HugeMatrix}
\description{
Implementation of triple matrix that can have indices that exceeds R's integer limit (2 ^ 31).
}
\details{
Set to overcome int32 limitation on indexation in R. \cr
The implementation is just triplet representation with a data.table. \cr
We use doubles to represent the indices. This limits the size of our new HugeMatrix to 2^53 - 1 \cr
}

