% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge_sparse.R
\name{HugeMatrix}
\alias{HugeMatrix}
\title{HugeMatrix}
\usage{
HugeMatrix(M, DM = mat.to.data.table(M), dims, dimnames)
}
\arguments{
\item{M}{Matrix to convert}

\item{DM}{Triplet form i, j, x needed}

\item{dims}{Dimension of the Matrix. Defaults to the dimension of M or the maximum of i and j}

\item{dimnames}{list with two entries, the first is the row names and the second the column names. Defaults to the dimension names of M}
}
\description{
HugeMatrix
}

