% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{indexation}
\alias{indexation}
\title{Indices of a Vector}
\usage{
indexation(X, N = length(X))
}
\arguments{
\item{X}{vector}

\item{N}{length of the vector}
}
\description{
Returns the indices of a vector. Either 1:N if N > 0 or integer(0).
}
