% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{inplace}
\alias{inplace}
\title{In place assignation}
\usage{
inplace(target, src)
}
\arguments{
\item{target}{List to modify}

\item{src}{Elements to assign to the list}
}
\description{
Modifies a list in place
}

