% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{insert}
\alias{insert}
\title{Inserting Data Into a Buffer}
\usage{
insert(EG, new.data, i, j, ...)
}
\arguments{
\item{EG}{The buffer}

\item{new.data}{data to be pushed to the buffer. When some columns in new.data are missing in the buffer, they will be added to the new buffer producing warnings.}

\item{i}{Row indices into which making the insertion}

\item{j}{Column indices into which making the insertion}
}
\description{
Inserting Data Into a Buffer
}

