% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.packages.R
\name{install.git}
\alias{install.git}
\title{Install Git}
\usage{
install.git(url, name, ...)
}
\arguments{
\item{url}{Remote url of the package}

\item{name}{Name of the package}
}
\description{
Installing package with git without using git2r package.
}

