% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.packages.R
\name{install.jalgos}
\alias{install.jalgos}
\title{Install From Jalgos Repos}
\usage{
install.jalgos(url, lib.loc = "lib", ...)
}
\description{
Wrapper around install_git to install packages from Jalgos repos (uses git2r)
}

