% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{is.positive}
\alias{is.positive}
\title{Testing Matrix Positivity}
\usage{
is.positive(S, tol = .Machine$double.eps, definite = FALSE)
}
\arguments{
\item{S}{Symmetric Matrix}

\item{tol}{Tolerance to decide whether an eigen value is null. Everything in the range [-tol, tol] is considered to be a null eigen value. Everything in ]-Inf, -tol[ is negative and everything in ]tol, Inf[ is positive.}

\item{definite}{Should we check the matrix be definite positive}
}
\description{
Tests whether a matrix is positive and optionally definite.
}

