% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{jacobi.inv}
\alias{jacobi.inv}
\title{Jacobi Inverse}
\usage{
jacobi.inv(a, b, p1 = Diagonal(x = 1/diag(a)), p = Diagonal(x = diag(a)),
  x = p1 \%*\% b, tol = sqrt(.Machine$double.eps), iter.max = 1000,
  logger = jlogger::JLoggerFactory("Jalgos Algebra"))
}
\arguments{
\item{a}{Matrix}

\item{b}{Right hand side of a.x = b}

\item{p1}{Predictioner, an approximation of the inverse of a. The closest it is from the real value the better. If it's too far away, it will not converge}

\item{p}{Inverse of the preconditioner.}

\item{x}{First guess of the solution}

\item{tol}{Convergence tolerance}

\item{iter.max}{Number maximum of iterations to achieve convergence}

\item{logger}{JLogger to log messages}
}
\description{
Simple implementation of the Jacobi Method
}
\references{
https://en.wikipedia.org/wiki/Jacobi_method
}
