% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{JF.S1}
\alias{JF.S1}
\title{Moments of a multidimensional normal variable}
\usage{
JF.S1(S, half.vec = TRUE, ...,
  logger = jlogger::JLoggerFactory("jalgos filter"))
}
\arguments{
\item{S}{Covariance matrix}

\item{half.vec}{Should the result concern only the free parameters of the distribution.}

\item{logger}{JLogger used to log messages}
}
\description{
Moments of a multidimensional normal variable
}
\details{
This function computes the 4th order moment of a multidimensional normal distribution of covariance matrix S. \cr
Let X be the random variable. This function computes the matrix E(tcrossprod(X %x% X))
}
