% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{js_versions}
\alias{js_versions}
\title{Get versions of bundled JavaScript libraries}
\usage{
js_versions()
}
\value{
A named list with the versions of Esprima, sass.js, Terser, and TypeScript.
}
\description{
This function returns the versions of the bundled JavaScript libraries used in the package.
}
\examples{
js_versions()
}
