% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{kronecker.proj}
\alias{kronecker.proj}
\title{Kronecker projection}
\usage{
kronecker.proj(M1, M2, Pl, Pr)
}
\arguments{
\item{M1}{a generic matrix that can be converted to a triplet using \code{mat.to.triplet}}

\item{M2}{DITTO}

\item{Pl}{A matrix that will project the rows of M1 \%x\% M2 into a space of smaller dimension. If not provided will assume identity}

\item{Pr}{A matrix that will project the columns of M1 \%x\% M2 into a space of smaller dimension. If not provided will assume identity}
}
\description{
Efficiently projects the result of a kronecker into two (or one) smaller spaces
}
