% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{load.or.install}
\alias{load.or.install}
\title{Installing packages}
\usage{
load.or.install(library.name, install.fun = install.packages, ...)
}
\arguments{
\item{library.name}{character string, name of the package to be installed}

\item{install.fun}{installation function to be used}

\item{...}{Parameters to be forwarded to the installation function}
}
\description{
Loads a package or install it if it's not installed yet
}

