% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{make.id}
\alias{make.id}
\title{Identify record}
\usage{
make.id(cols, vals)
}
\arguments{
\item{cols}{Column names}

\item{vals}{Values}
}
\description{
Creates a column that uniquely identifies a set of records
}
