% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.vectorization.R
\name{mat.index}
\alias{mat.index}
\alias{mat.index}
\alias{index.sym}
\alias{reverse.index.sym}
\alias{assign.matrix.vect}
\title{Matrix index to vectorization index}
\usage{
mat.index(i, j, n)

index.sym(i, j, n, keep.diag = TRUE)

reverse.index.sym(I, n, keep.diag = TRUE)

assign.matrix.vect(M, i, j, values)
}
\arguments{
\item{i}{row index}

\item{j}{column index}

\item{n}{number of rows}

\item{keep.diag}{is the diagonal kept in the half vectorization}

\item{I}{Half vectorization index}

\item{M}{matrix}

\item{values}{numeric vector}
}
\section{Functions}{
\itemize{
\item \code{mat.index}: Computes the vectorization index corresponding to a given row column combination in a matrix

\item \code{index.sym}: Computes the half vectorization index corresponding to a given row column combination in a matrix

\item \code{reverse.index.sym}: Computes the row index and column index corresponding to a given vectorization index

\item \code{assign.matrix.vect}: Assigns value to a matrix by a triplet specification
}}

