% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge_sparse.R
\name{mat.join}
\alias{mat.join}
\title{Matrix Merger}
\usage{
mat.join(Dx, Dy, ...)
}
\arguments{
\item{Dx}{LHS buffer}

\item{Dy}{RHS buffer}
}
\description{
Function that will join two triplet representation to compute the matrix product
}

