% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huge_sparse.R
\name{mat.mult}
\alias{mat.mult}
\title{Generic Matrix Multiplication}
\usage{
mat.mult(x, y, ...)
}
\arguments{
\item{x}{RHS Matrix}

\item{y}{LHS Matrix}

\item{...}{Additional settings}
}
\description{
Generic function that will implement matrix multiplication as well as \code{crossprod} and and \code{tcrossprod}
}

