% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\docType{methods}
\name{mat.to.data.table}
\alias{mat.to.data.table}
\alias{mat.to.data.table,Matrix-method}
\alias{mat.to.data.table,diagonalMatrix-method}
\alias{mat.to.data.table,matrix-method}
\alias{mat.to.data.table,symmetricMatrix-method}
\alias{mat.to.data.table,vector-method}
\alias{mat.to.triplet}
\title{Matrix to data table}
\usage{
mat.to.data.table(M, ...)

\S4method{mat.to.data.table}{Matrix}(M, one.based = TRUE,
  with.names = FALSE, ...)

\S4method{mat.to.data.table}{matrix}(M, one.based = TRUE,
  with.names = FALSE, ...)

\S4method{mat.to.data.table}{diagonalMatrix}(M, one.based = TRUE, ...)

\S4method{mat.to.data.table}{symmetricMatrix}(M, one.based = TRUE, ...)

\S4method{mat.to.data.table}{vector}(M, one.based = TRUE,
  with.names = FALSE, ...)
}
\arguments{
\item{M}{Matrix to transform to its triplet form}

\item{one.based}{Indices start at 1 or 0}

\item{with.names}{Should rownames and column names be added to the data.table}
}
\description{
Transforms a matrix into its triplet version in a data.table
}

