% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_regex.R
\name{match.grep}
\alias{match.grep}
\title{Retrieving an element of a list by matching the names}
\usage{
match.grep(L, name)
}
\arguments{
\item{L}{Named list. The names are regexes that attempt to match \code{name}.}

\item{name}{string to be matched}
}
\value{
The first occurence of L for which its name matches the parameter \code{name}
}
\description{
Retrieving an element of a list by matching the names
}

