% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{matching.cols}
\alias{matching.cols}
\title{Matching columns}
\usage{
matching.cols(DT, match.exp, ig.match.exp = "^$")
}
\arguments{
\item{DT}{a data table}

\item{match.exp}{the regular expression to be matched by column names}

\item{ig.match.exp}{the regular expression to be unmatched by column names}
}
\description{
Retrieve the column names that match a pattern AND does not match an other pattern
}
