% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{matrix.norm}
\alias{matrix.norm}
\alias{matrix.norm}
\alias{matrix.norm.inf}
\alias{matrix.scal}
\title{Norms of a Matrix}
\usage{
matrix.norm(M)

matrix.norm.inf(M)

matrix.scal(M1, M2, normalized = TRUE)
}
\arguments{
\item{M}{a matrix}

\item{M1}{a matrix}

\item{M2}{a matrix}

\item{normalized}{Should the scalar product be normalized to be expressed as a fraction of the norm of 'M1'.}
}
\details{
Simple wrappers around the norm function.
}
\section{Functions}{
\itemize{
\item \code{matrix.norm}: L2 norm (frobenius norm)

\item \code{matrix.norm.inf}: Infinite norm (max of the absolute value of the coefficients)

\item \code{matrix.scal}: scalar product of matrices. Identitical to the scalar product of the vectorized form of the matrices
}}

