% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{matrix.scal.raw}
\alias{matrix.scal.raw}
\title{Scalar Product Matrix}
\usage{
matrix.scal.raw(M1, M2)
}
\arguments{
\item{M1}{Matrix}

\item{M2}{Matrix}
}
\description{
Computes the scalar product between two matrices. Simply the sum of the product Hadamard product
}
