% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{mem.safe.kronecker}
\alias{mem.safe.kronecker}
\title{Memory Safe Kronecker Product}
\usage{
mem.safe.kronecker(S1, S2, krsize.limit = krmem.limit * 2^26,
  krmem.limit = Inf, keep.diag = FALSE, ..., logger = NULL)
}
\arguments{
\item{S1}{First matrix}

\item{S2}{Second matrix}

\item{krsize.limit}{The limit in number of non zero value of the result}

\item{krmem.limit}{Used to express the limit in GBs}

\item{keep.diag}{Should the diagonal terms be kept}

\item{...}{No use now}

\item{logger}{JLogger to log additional messages}
}
\description{
Performs a kronecker product of two so the memory used stays under a scpecified bound.
}
\details{
The matrices are divided into smaller chunks and the kronecker products of these chunks are computed. Only the most significant correlation are kept to meet the target size.
The approximation error should be small if there are a lot fo values < 0.01 in the correlation matrix. \cr
The memory taken by a Matrix (in the triplet form) with N non zero elements is roughly 16 * N / 2 ^ 20 MBs (two integer vectors and a double vector: 4 + 4 + 8) \cr
If you want to limit the memory to 1Gb you need to set the size.limit to 2 ^ 26. \cr
Keep in mind however that the chunking process will use twice the size provided. If you want to limit the mmeory usage to 2Gb give 1Gb as parameter.
}
