% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{memory.usage}
\alias{memory.usage}
\alias{memory.usage}
\alias{mb.object.size}
\title{Memory usage}
\usage{
memory.usage(LOBJS = NULL, envir = .GlobalEnv, all = TRUE, nshow = 10)

mb.object.size(x)
}
\arguments{
\item{LOBJS}{List of objects to evaluate}

\item{envir}{Environment to evaluate}

\item{all}{Should all objects be considered}

\item{nshow}{Only show the nshow biggest object}

\item{x}{the object}
}
\description{
Tools to monitor memory usage
}
\section{Functions}{
\itemize{
\item \code{memory.usage}: Outputs the memory usage of all R object in a given environment

\item \code{mb.object.size}: Gets the size of an object in MBs
}}

