% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{nearly.equal}
\alias{nearly.equal}
\alias{nearly.in.set}
\title{Equality for doubles}
\usage{
nearly.equal(x, y, tol = .Machine$double.eps^0.5)

nearly.in.set(x, set, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{tol}{Tolerance for equality i.e abs(x - y) < tol}

\item{set}{numeric vector}
}
\description{
Checks for equality allowing for a margin of error
}
\section{Functions}{
\itemize{
\item \code{nearly.equal}: Check if two doubles are equal up to a given margin of error

\item \code{nearly.in.set}: Check if a double are equal up to a given margin of error
}}

