% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.non.finite.R
\name{non.finite}
\alias{non.finite}
\alias{outlier.filter}
\alias{replaceNA}
\alias{safe.cor}
\alias{safe.max}
\alias{safe.mean}
\alias{safe.min}
\alias{safe.sum}
\alias{safe.var}
\title{Handling non finite values}
\usage{
safe.sum(X, Y, finite.fun = is.finite, replacement.value = 0)

safe.mean(..., finite.fun = is.finite)

safe.max(X, ...)

safe.min(X, ...)

safe.var(X, Y = X, var.fun = var, ...)

safe.cor(X, Y, ...)

replaceNA(X, Y = 0, finite.fun = is.finite)

outlier.filter(X, qt = 0.01, down.qt = qt/2, up.qt = 1 - qt/2, ...)
}
\arguments{
\item{X}{Numeric vector}

\item{Y}{Numeric vector}

\item{finite.fun}{Function that return TRUE if a value is finite}

\item{...}{Numeric vectors of equal length}

\item{var.fun}{Variance function to be used}

\item{qt}{Proportion of the distribution to filter out. Removes the same amount in both ends.}

\item{down.qt}{Proportion to ommit in the lower tail}

\item{up.qt}{Proportion to ommit in the upper tail}
}
\section{Functions}{
\itemize{
\item \code{safe.sum}: Sum where non finite values are replaced with 0. Returns X + Y where non finite values in both vectors are replaced with 0

\item \code{safe.mean}: Averages several vectors ommiting non finite values

\item \code{safe.max}: Gets the max of a numeric vector without warnings

\item \code{safe.min}: Gets the min of a numeric vector without warnings

\item \code{safe.var}: Computes the variance on the filtered data

\item \code{safe.cor}: Computes the correlation on the filtered data

\item \code{replaceNA}: Replace non finite entries

\item \code{outlier.filter}: Returns a logical vector that filters out data outside of the range defined by the given quantile
}}

