% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{not.null.names}
\alias{not.null.names}
\title{Alternative To NULL Dimensions}
\usage{
not.null.names(x, name.fun = names, length.fun = length, default = "")
}
\arguments{
\item{x}{The object on which name.fun and length.fun are to be applied}

\item{name.fun}{function returning the desired name of x}

\item{length.fun}{function returning the length of the desired dimension}
}
\description{
Returns a vector of empty strings instead of NULL
}
