% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{orthonorm}
\alias{orthonorm}
\title{Orthonormalizing a Matrix}
\usage{
orthonorm(B)
}
\arguments{
\item{B}{base to orthonormalize}
}
\description{
Returns the closest orthonormal matrix using the function \code{qr}
}
\seealso{
\link{qr}, \link{qr.Q}
}
