% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applies.R
\name{over}
\alias{over}
\title{Over}
\usage{
over(L, FUN, default.value = NULL, ...)
}
\arguments{
\item{L}{List of objects}

\item{FUN}{Function to be recursively applied}

\item{default.value}{The result of the function in case the list provided is empty}
}
\value{
FUN(X[1], FUN(X[2], FUN(. . ., FUN(X[N-1], X[N]), `...`), . . .), `...`), `...`). If the list has only one element it will return X[1]
}
\description{
Applies a bivariate function recursively to a list and the result of the previous call.
}
\details{
The first to argument of the function should be of type of the object in the list provided.
}
\examples{
\code{D = data.table(x = 1:10, u = letters[1:10])}
\code{over(list(D, D, D, D), rbind)}
}
