% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.conversion.R
\name{partial.conversion}
\alias{partial.conversion}
\title{Type Conversion}
\usage{
partial.conversion(conv.fun, default.val, null.val = "")
}
\arguments{
\item{conv.fun}{Conversion function}

\item{default.val}{Default value}

\item{null.val}{A vector of values that should not be converted}
}
\description{
Converts object that don't match a set of values
}
