% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{plus.join}
\alias{plus.join}
\title{Plus Join}
\usage{
plus.join(D1, D2, col.plus = setdiff(intersect(names(D1), names(D2)), by),
  all = TRUE, ...)
}
\arguments{
\item{D1}{first data.table}

\item{D2}{second data.table}

\item{col.plus}{Columns to add}

\item{all}{Forwarded to data.table::merge. Control which records should be kept}

\item{...}{Parameters forwarded to data.table::merge. 'by' needs to be provided}
}
\description{
Plus join is a type of join where the columns of the same name are added together. Missing values in one of the two tables are considered to be 0 by default
}

