% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.control.flow.R
\name{print.call.stack}
\alias{print.call.stack}
\title{Print Call Stack}
\usage{
print.call.stack(..., logger = NULL)
}
\arguments{
\item{...}{not used}
}
\description{
Prints the entire call stack. Meant to be called upon an error occurring. Especially handy for the pbdR context where 'options(error = browser)' is unavailable.
}

