% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{psolve}
\alias{psolve}
\title{Partial Inversion Using Singular Values}
\usage{
psolve(a, b, var.thresh = 0.95)
}
\arguments{
\item{a}{Matrix to invert}

\item{b}{RHS of the equation a.v = b where v is the unknown}

\item{var.thresh}{quantity of volatility to be taken into account for inverstion}
}
\description{
Inverts a matrix by truncating the eigen values that contribute to 95% of the whole volatility.
}

