% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{put.names}
\alias{put.names}
\title{Names a list}
\usage{
put.names(L, name = "x", ...)
}
\arguments{
\item{L}{Either a list or a vector that will be put in a singleton}

\item{name}{Name to be used}
}
\description{
Names a list
}
