% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{random.cat.var}
\alias{random.cat.var}
\title{Random categorial}
\usage{
random.cat.var(var, set.size)
}
\arguments{
\item{var}{List specifying the distribution. Fields 'name' is mandatory, then either a list of items in the field 'items' or a number of items in the fiels 'nvars' need to be specified. Finally the field 'replace' controls whether items can be drawn several time. If missing defaults to TRUE}

\item{set.size}{Size of the sample to be drawn}
}
\description{
Draws an arbitrary number of items from a categorical distribution
}
\seealso{
random.data.set
}
