% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{random.data.set}
\alias{random.data.set}
\title{Random data set}
\usage{
random.data.set(set.size, varsdescription)
}
\arguments{
\item{set.size}{Size of the data set}

\item{varsdescription}{A list of categorical variable description to be forwarded to random.cat.var}
}
\description{
Generate a random data sets using random.cat.var
}
\seealso{
random.cat.var
}
