% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.random.R
\name{random.date}
\alias{random.date}
\title{Random Dates}
\usage{
random.date(n, date.ref = as.POSIXct(Sys.time()), rdist = function(..., sd =
  1e+07) rnorm(..., sd = sd), ...)
}
\arguments{
\item{n}{Sample size}

\item{date.ref}{Reference date}

\item{rdist}{distribution function to be used}

\item{...}{To be forwarded to `rdist`}
}
\description{
Generate random dates around a reference date
}

