% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.matrices.R
\name{random.mat}
\alias{random.mat}
\alias{random.matrix}
\alias{random.sym.matrix}
\alias{random.anti.sym.matrix}
\alias{random.cov.matrix}
\alias{random.sparse.matrix}
\alias{random.sparse.cov.matrix}
\alias{random.orthonormal.matrix}
\alias{random.orthogonal}
\title{Generating random matrices}
\usage{
random.matrix(n, p = n, rgen = rnorm, ...)

random.sym.matrix(n, rgen = rnorm, ...)

random.anti.sym.matrix(n, rgen = rnorm, ...)

random.cov.matrix(n, tot.vol = n, ...)

random.sparse.matrix(n, p = n, N, rgen = rnorm)

random.sparse.cov.matrix(N, ...)

random.orthonormal.matrix(n)

random.orthogonal(p, q = p, vals = 1 + 1:p)
}
\arguments{
\item{n}{Number of rows}

\item{p}{number of factors}

\item{rgen}{Distribution to draw from. Defaults to a normal with 0 mean and variance 1}

\item{...}{Parameters to be passed to the random number generator and to lower level functions}

\item{tot.vol}{Scaling that control the total volatiliy of the covariance matrix created}

\item{q}{number of variables}

\item{vals}{Values for each factor}
}
\description{
Generating random matrices
}
\section{Functions}{
\itemize{
\item \code{random.matrix}: Generates a random matrix of independently drawn realisations of a random variable

\item \code{random.sym.matrix}: Generates a random symmetric matrix of independently drawn realisations of a random variable

\item \code{random.anti.sym.matrix}: Generates a random anti symmetric matrix of independently drawn realisations of a random variable

\item \code{random.cov.matrix}: Generates a random covariance matrix. Uses random.sym.matrix

\item \code{random.sparse.matrix}: Creates a random sparse matrix

\item \code{random.sparse.cov.matrix}: Creates a random sparse covariance matrix. Sparsity is hard to control exactly. The target is approximate. Uses random.sparse.matrix

\item \code{random.orthonormal.matrix}: Creates a random orthonormal.matrix. Effectively draws an anti symmetric matrix and uses the Cayley parametrization to create the member of SOn(R)

\item \code{random.orthogonal}: Creates a random orthogonal matrix.
}}

