% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.random.R
\name{random.word}
\alias{random.word}
\title{Random Words}
\usage{
random.word(n, wsize = 100, wprob = 0.1)
}
\arguments{
\item{n}{Number of words to be generated}

\item{wsize}{parameter `size` of the binomial sample to create word lengths}

\item{wprob}{parameter `prob` of the binomial sample to create word lengths}
}
\description{
Generate an arbitrary number of random strings made of letters from the 26 characters alphabet.
}
