% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{ranking.rating}
\alias{ranking.rating}
\alias{rating.normalize}
\alias{rank.normalize}
\title{Rankings ratings}
\usage{
rating.normalize(X, fun.proj = exp, filter = TRUE)

rank.normalize(X, rank.norm)
}
\arguments{
\item{X}{Numeric vector}

\item{fun.proj}{Function to project ratings to a different scale}

\item{filter}{Subset to the ratings to take the max from}

\item{rank.norm}{Rank to use to recenter ratings}
}
\details{
Tools to rescale and recenter ratings. It is usually useful to rescale ratings according to the maximum value that occurs or according to the nth biggest value
}
\section{Functions}{
\itemize{
\item \code{rating.normalize}: Applies a transformation and rescale the rankings so they now appear as a proportion of the biggest value

\item \code{rank.normalize}: Recenter values so the a given rank has now a value of 0
}}

