% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{rbind.apply}
\alias{rbind.apply}
\title{rbind apply}
\usage{
\method{rbind}{apply}(vect, fun, fill = FALSE, ...)
}
\arguments{
\item{vect}{apply function on this vector}

\item{fun}{function to apply on vect. Should return a data.table}

\item{fill}{rbind argument}
}
\description{
equivalent to rbindlist(lapply(...))
}
