% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{recursive.paste}
\alias{recursive.paste}
\alias{mpaste}
\alias{mpaste0}
\title{Recursive paste}
\usage{
mpaste(c1, c2, ...)

mpaste0(c1, c2, ...)
}
\arguments{
\item{c1}{Character vector}

\item{c2}{Character vector}

\item{...}{parameters to be forwarded to paste or paste0}
}
\description{
Computes all the possible string concatenation between members of c1 and c2
}
\section{Functions}{
\itemize{
\item \code{mpaste}: Uses paste

\item \code{mpaste0}: Uses paste0
}}

\seealso{
paste, paste0
}
