% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_regex.R
\name{replace.whole.string}
\alias{replace.whole.string}
\title{describeIn text.to.data Replaces the entirety of a string with the replacement pattern. Useful when we are more concerned with discarding everything that doesn't match than the contrary.}
\usage{
replace.whole.string(pattern, replacement, text, ...)
}
\description{
describeIn text.to.data Replaces the entirety of a string with the replacement pattern. Useful when we are more concerned with discarding everything that doesn't match than the contrary.
}
