% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.packages.R
\name{require.or.install}
\alias{require.or.install}
\alias{require.or.install.dev}
\title{Install Missing Package}
\usage{
require.or.install(
  name,
  install.fun = install.git,
  vars = c(`$group` = group, `$name` = name),
  ...,
  force = FALSE
)

require.or.install.dev(name, url.pattern = file.package.pattern, ...)
}
\arguments{
\item{name}{Name of the package}

\item{install.fun}{Package installation function to use}

\item{vars}{Variables that appear in the url.pattern}

\item{...}{to be forwarded to install.fun}

\item{force}{force install}

\item{url.pattern}{Url pattern}

\item{group}{Group to which the package belongs}

\item{url}{Url of the package}
}
\description{
Install Missing Package
}
\section{Functions}{
\itemize{
\item \code{require.or.install}: Install specified package in case call to \code{require} fails.

\item \code{require.or.install.dev}: For development use. Installs package directly from directory not from remote repo
}}

