% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{reserve}
\alias{reserve}
\title{Resize, reserve, binds}
\usage{
reserve(M, nrows, ncols, padding.val = 0, zero.based = FALSE, ...)
}
\description{
Reserves additional data, so when more data is inserted no copy is dones
}
\details{
Some objects are meant to be growing as some algorithms run. These functions provides feature to handle these cases
}
