% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{resize}
\alias{resize}
\title{Resize, reserve, binds}
\usage{
resize(M, nrows, ncols, padding.val = 0, zero.based = FALSE, ...)
}
\description{
Resizes a buffer. If reserve was called earlier no copy should be made
}
\details{
Some objects are meant to be growing as some algorithms run. These functions provides feature to handle these cases
}
