% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{resize.reserve}
\alias{resize.reserve}
\title{Resizing, reserving}
\description{
Resizing, reserving
}
\details{
Some objects are meant to be growing as some algorithms run. These functions provides feature to handle these cases
}
