% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{roll.merge.2d}
\alias{roll.merge.2d}
\title{Roll merge in 2D}
\usage{
roll.merge.2d(DT1, DT2, cols1, cols2, grid.size, ...)
}
\arguments{
\item{DT1}{data.table}

\item{DT2}{data.table}

\item{cols1}{2 column names to use for the merge in DT1}

\item{cols2}{2 column names to use for the merge in DT2}

\item{grid.size}{aggregation size. A big value reduce the accuracy of the merge}
}
\description{
Divide the tables in a 2D grid, and match the points that are in the closest grid cells. While not all the points in DT1 are matched with a point in DT2, keep looking in further cells
}
