% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{rotation}
\alias{rotate.vector}
\alias{rotation}
\alias{rotation.matrix}
\title{Rotation}
\usage{
rotation.matrix(theta)

rotate.vector(X, theta)
}
\arguments{
\item{theta}{angle of the rotation}
}
\value{
A 2 x 2 matrix that when multiplied to a vector produce the rotated vector
}
\section{Functions}{
\itemize{
\item \code{rotation.matrix}: Computes a two dimensional rotation matrix

\item \code{rotate.vector}: Rotates a two dimensional vector of the specified angle
}}

