% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{row.bind}
\alias{row.bind}
\title{Resize, reserve, binds}
\usage{
row.bind(M1, M2)
}
\description{
Binds two data structures so that the number of rows of the result is the sum of the number of rows of the two arguments. Columns have to match.
}
\details{
Some objects are meant to be growing as some algorithms run. These functions provides feature to handle these cases
}
