% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{rows.cols}
\alias{rows.cols}
\alias{rows}
\alias{cols}
\title{Row numbers and column numbers}
\usage{
rows(M)

cols(M)
}
\section{Functions}{
\itemize{
\item \code{rows}: returns a vector containing all the row nmbers

\item \code{cols}: returns a vector containing all the column numbers
}}

