% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{safe.cov2cor}
\alias{safe.cov2cor}
\title{Covariance Matrix To Correlation Matrix}
\usage{
safe.cov2cor(M, ...)
}
\arguments{
\item{M}{Matrix to transform}

\item{sqDl}{Diagonal to multiply on the left side}

\item{sqDr}{Diagonal to multiply on the right side}

\item{diag.fun}{Function to generate the diagonal matrix}
}
\description{
Transforms a covariance matrix into a correlation matrix with parameters to handle
}
