% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.S4.R
\name{safeGeneric}
\alias{safeGeneric}
\title{Not overwriting generics}
\usage{
safeGeneric(f, ...)
}
\arguments{
\item{f}{Function to be made generic}

\item{...}{Extra parameters to be passed on to setGeneric}
}
\description{
Creates a generic only if f is not a generic yet. Useful when generics are created in different files and the user doesn't have control over the order in which it's loaded
}
