% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{scale.center}
\alias{center}
\alias{scale}
\alias{scale.and.center}
\alias{scale.center}
\title{Scaling and centering numeric vectors}
\usage{
center(X, cent = mean(X, na.rm = TRUE))

scale(X, scaling = sd(X, na.rm = TRUE))

scale.and.center(X, cent = mean(X, na.rm = TRUE), scaling = sd(X, na.rm =
  TRUE))
}
\arguments{
\item{X}{numeric vector}

\item{cent}{Centering}

\item{scaling}{Scaling}
}
\description{
tools to center numeric vectors around their mean and scale them by their standard deviation
}
\section{Functions}{
\itemize{
\item \code{center}: Centers a numeric vector

\item \code{scale}: Scales a numeric vector

\item \code{scale.and.center}: Center then scales
}}

