% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{sdef.solve}
\alias{sdef.solve}
\title{Semi Definite Positive Inverse}
\usage{
sdef.solve(a, b, val.rel.tol)
}
\arguments{
\item{a}{Matrix to invert}

\item{b}{RHS of the equation a.v = b where v is the unknown}
}
\description{
Inverts a semi definite positive matrix
}
