% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{set.convention.names}
\alias{set.convention.names}
\title{Set convention names}
\usage{
set.convention.names(
  DT,
  sep = "_",
  unwanted.chars = c(é = "e", è = "e"),
  logger = JLoggerFactory("Jalgos Utils")
)
}
\arguments{
\item{DT}{data.table}
}
\description{
Turn every column name of a data table to a convention format with lower case and point separated column names
}
