% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.vectorization.R
\name{set.distribution}
\alias{set.distribution}
\alias{mat.vectorization.distributed}
\alias{switch.local.distributed}
\title{Set Distribution}
\usage{
mat.vectorization.distributed(new.context)

switch.local.distributed(fun.local, fun.distributed)
}
\arguments{
\item{new.context}{New context to set (TRUE or FALSE) if not provided returns the current context}
}
\description{
Set Distribution
}
\section{Functions}{
\itemize{
\item \code{mat.vectorization.distributed}: Gets/ sets the distribution behaviour of the matrix vectorization function

\item \code{switch.local.distributed}: returns the proper function given the distributed context
}}

