% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.vectorization.R
\name{set.mat.vectorization.distributed}
\alias{set.mat.vectorization.distributed}
\title{Set Distribution}
\usage{
set.mat.vectorization.distributed(new.context)
}
\arguments{
\item{new.context}{New context to set (TRUE or FALSE) if not provided returns the current context}
}
\description{
Gets/ sets the distribution behaviour of the matrix vectorization function
}
