% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{simple.function}
\alias{const.fun}
\alias{null.fun}
\alias{one.const.fun}
\alias{simple.function}
\title{Simple functions}
\usage{
null.fun(...)

one.const.fun(...)

const.fun(..., val = 0)
}
\description{
Aliases for very basic function that are used here and there
}
\section{Functions}{
\itemize{
\item \code{null.fun}: returns 0

\item \code{one.const.fun}: returns 1

\item \code{const.fun}: returns \code{val}
}}

