% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{sparsity.ratio}
\alias{sparsity.ratio}
\title{Sparsity Ratio}
\usage{
sparsity.ratio(M)
}
\arguments{
\item{M}{a matrix}
}
\description{
Computes the ratio between the non empty elements and the total elements of a matrix
}
