% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.R
\name{split.write.csv}
\alias{split.write.csv}
\title{Split Write csv}
\usage{
\method{split}{write.csv}(TAB, split.size, logger = get.grt.logger())
}
\arguments{
\item{TAB}{parameter Data.table, usually big, to be splitted and exported as csv files}

\item{split.size}{parameter controlling the max size of splits, in number of lines}
}
\description{
Generate an arbitrary number of random strings made of letters from the 26 characters alphabet.
}

