% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{sum.non.finite}
\alias{sum.non.finite}
\title{Non Finite Coefficients}
\usage{
sum.non.finite(M)
}
\arguments{
\item{M}{the matrix}
}
\description{
Returns the number of non finite entries in a matrix
}

