% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{to.local}
\alias{to.local}
\title{Global To Local Buffer}
\usage{
to.local(obj)
}
\arguments{
\item{obj}{object to gather}
}
\description{
Transform a global Buffer into a local buffer to process 0. Mostly for testing purposes.
}
