% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.control.flow.R
\name{trace.upon.error}
\alias{trace.upon.error}
\title{Print Call Stack Upon Error}
\usage{
trace.upon.error(logger)
}
\arguments{
\item{logger}{The logger that will print out the stack}
}
\description{
Prints the call stack and quits if the session is not interactive, sets option error to `recover` otherwise.
}
