% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.S4.R
\name{traits.dispatch}
\alias{traits.dispatch}
\alias{traits.apply}
\alias{double.traits.apply}
\title{Applying Sub Methods For Traits}
\usage{
traits.apply(OBJ = NULL,
                      method,
                      traits.traj,
                      ...,
                      update.var = !is.null(OBJ),
                      until.no.null = FALSE,
                      prev.objs = list(),
                      post.objs = list(),
                      tra.logger = JF.logger())

double.traits.apply(
  method,
  ...,
  traits.traj1 = get.direct.superclasses(largs[[obj1]]),
  traits.traj2 = get.direct.superclasses(largs[[obj2]]),
  obj1,
  obj2
)
}
\arguments{
\item{OBJ}{To be used in case each call to the submethods is supposed to transform the input. In this case OBJ must be set to the object on which the dispatch occurs}

\item{method}{The method that we want to dispatch}

\item{traits.traj}{The traits that constitute the given object}

\item{...}{Parameters of the submethods}

\item{update.var}{Should we update the object after each sub call. Defaults to whether OBJ was provided}

\item{until.no.null}{Should the function return as soon as a result is not null?}

\item{prev.objs}{To complement the signature. Objects that appear before the concerned class}

\item{post.objs}{To complement the signature. Objects that appear after the concerned class}

\item{tra.logger}{Logger that helps trace the dispatching}

\item{traits.traj1}{Traits of obj1}

\item{traits.traj2}{Traits of obj2}

\item{obj1}{The first argument to dispatch on}

\item{obj2}{The second argumet to dispatch on}
}
\description{
Applying Sub Methods For Traits
}
\section{Functions}{
\itemize{
\item \code{traits.apply}: Applies the sub methods for each sub traits of a given class

\item \code{double.traits.apply}: Dispatches a method when both objects are a composition of several traits and the sub methods need to be called on each combination of traits
}}

