% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{trim.cov.matrix}
\alias{trim.cov.matrix}
\title{Removing Least Significant Covariance Coefficients}
\usage{
trim.cov.matrix(S, sqDl = sqrt(posD(diag(M))), sqDr = sqDl, ...,
  mtype = "", tol = 0, tol.inv = 10 * .Machine$double.eps,
  logger = NULL)
}
\arguments{
\item{S}{The covariance matrix}
}
\description{
Will reduce the size of a covariance matrix by removing its least significant coefficients. The trimming will happen in the correlation matrix in order not to overweight the entries with the biggest variance.
}

