% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{triplet.product}
\alias{triplet.product}
\alias{triplet.prod.rb}
\alias{triplet.prod.hash}
\title{Triplet Product}
\usage{
triplet.prod.rb(i1, j1, x1, i2, j2, x2)

triplet.prod.hash(i1, j1, x1, i2, j2, x2)
}
\arguments{
\item{i1}{row indices in LHS Matrix}

\item{j1}{column indices in LHS Matrix}

\item{x1}{values in LHS Matrix}

\item{i2}{row indices in RHS Matrix}

\item{j2}{column indices in RHS Matrix}

\item{x2}{values in RHS Matrix}
}
\details{
Two implementation of matrix products that uses maps (red / black tree and hashmap) to perform a matrix multiplication. This is a more efficient implementation for very sparse huge matrices. \cr
The hashmap version is almost always preferrable.
}
\section{Functions}{
\itemize{
\item \code{triplet.prod.rb}: Uses a read / black tree map to represent matrix data.

\item \code{triplet.prod.hash}: Uses a hashmap to represent the data
}}

