% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.vectorization.R
\name{tripleting}
\alias{tripleting}
\alias{vech.triplet}
\title{Triplet to (half) vectorization}
\usage{
vech.triplet(DM = mat.to.triplet(M), nr = nrow(M), M, keep.diag = TRUE)
}
\arguments{
\item{DM}{Triplet representation i, j, x in a data.table}

\item{nr}{Number of rows of the matrix}

\item{M}{the matrix of which DM is the triplet representation}

\item{keep.diag}{should the diagonal be kept in the half vectorization}
}
\section{Functions}{
\itemize{
\item \code{vech.triplet}: Transforms a triplet representation of a matrix into its half vectorization
}}

