% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{try.catch}
\alias{try.catch}
\title{Try Catch With Switch To Enable / Disable}
\usage{
try.catch(expr, ..., dev.mode = FALSE)
}
\arguments{
\item{expr}{Expression to execute}

\item{...}{Arguments to be forwarded to tryCatch}

\item{dev.mode}{If set to true won't try to catch any error}
}
\description{
One may want to disable Try / Catch to see what causes an error. This function is a simple wrapper around \code{tryCatch} to do so.
}
