% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typescript.R
\name{typescript_transpile}
\alias{typescript_transpile}
\title{Transpile TypeScript code to JavaScript}
\usage{
typescript_transpile(input, transpileOptions = list())
}
\arguments{
\item{input}{A character string containing the TypeScript code to be transpiled.}

\item{transpileOptions}{A list of options to pass to the TypeScript transpiler. See the
\href{https://www.typescriptlang.org/docs/handbook/compiler-options.html}{TypeScript documentation} for available options.}
}
\value{
A list containing the transpiled JavaScript code and any diagnostics.
}
\description{
This function uses the TypeScript compiler to transpile TypeScript code into JavaScript.
}
\examples{
ts_code <- "const greet = (name: string): string => `Hello, ${name}!`;"
typescript_transpile(ts_code, list(compilerOptions = list(target = "ES5")))

}
