% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{update.buffer}
\alias{update.buffer}
\title{Replacing data in the buffer}
\usage{
update.buffer(EG, new.data, ...)
}
\arguments{
\item{EG}{The buffer}

\item{new.data}{Data to use as replacement}
}
\description{
Replaces parts or totality of the buffer.
}
