% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.S4.R
\name{update.obj}
\alias{update.obj}
\title{Updating object}
\usage{
update.obj(obj, ...)
}
\arguments{
\item{obj}{Object to update}

\item{...}{new fields of the object. Named arguments update slots and unnamed arguments update parent object.}
}
\description{
Creates a copy of the object which slots and parents can be updated via new.
}
\details{
S4 classes slots cannot be modified. Solution such as \code{obj@slot <- new.value} are discouraged as the operation creates a copy of obj everytime.\cr
The proper way to update an S4 object is to use the \code{new} operator. The problem with new is that it requires the class name. Hardcoding it prevents sub classes to use a parent function properly.\cr
The solution is to give the class of the object to be updated as a parameter of new. To be DRY we provide this function.
}
