% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{user.io}
\alias{user.io}
\alias{get.keyboard.input}
\alias{press.enter.to.continue}
\alias{y.to.continue}
\title{Reading input from user}
\usage{
get.keyboard.input(...)

press.enter.to.continue(...)

y.to.continue(yes.value = "^yes$", ...)
}
\arguments{
\item{...}{Forwarded to grepl}

\item{yes.value}{If the user input matches this pattern the function returns TRUE}
}
\details{
User input terminates when the user presses 'Enter'
}
\section{Functions}{
\itemize{
\item \code{get.keyboard.input}: Suspends execution and return the user input

\item \code{press.enter.to.continue}: Suspends execution and prompts the user to press 'Enter' to continue

\item \code{y.to.continue}: Suspends execution. User must enter a pattern matching yes.value to continue
}}

