% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.vectorization.R
\name{vec.reverse}
\alias{vec.reverse}
\title{Full vectorization}
\usage{
vec.reverse(V, n = findN(vecdim(V), FALSE), p = n, ...,
  mat.constr = safe.mat.constr)
}
\description{
Builds the matrix corresponding to a vectorization
}
