% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{vecdim}
\alias{vecdim}
\title{Get The Dimension Of an Algebraic Vector}
\usage{
vecdim(x)
}
\arguments{
\item{x}{vector}
}
\description{
Returns the dimension of an algebraic vector. Will work with objects of type 'vector' and one row matrices
}
