% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.vectorization.R
\docType{methods}
\name{vech}
\alias{vech}
\alias{vech,ANY-method}
\title{Half vectorization}
\usage{
vech(M, ...)

\S4method{vech}{ANY}(M, keep.diag = TRUE)
}
\description{
Performs the half vectorization of a symmetric matrix.
}
\details{
Half vectorization can be done in a pretty straightforward fashion for sparse matrices. The solution here preserves sparsity which makes sure that the program will not run out of memory because of an unwanted conversion to a dense matrix
}
\references{
https://en.wikipedia.org/wiki/Vectorization_\%28mathematics\%29#Half-vectorization
}

