% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.vectorization.R
\docType{methods}
\name{vech.reverse}
\alias{vech.reverse}
\alias{vech.reverse,ANY-method}
\title{Reverse half vectorization}
\usage{
vech.reverse(V, ...)

\S4method{vech.reverse}{ANY}(V, keep.diag = TRUE, symmetric = TRUE,
  mat.constr = safe.mat.constr)
}
\arguments{
\item{V}{Half vectorization of the matrix}

\item{keep.diag}{Was the diagonal kept when performing the half vectorization}

\item{symmetric}{Is the matrix symmetric. If not then the matrix is assumed antisymmetric}
}
\description{
Take an half vectorization and returns the corresponding matrix.
}

