% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.vectorization.R
\name{vech.reverse.triplet}
\alias{vech.reverse.triplet}
\title{Half vectorization triplet to matrix triplet}
\usage{
vech.reverse.triplet(
  DV = mat.to.triplet(V),
  nr = findN(nrow(V)) + !keep.diag,
  V,
  keep.diag = TRUE
)
}
\arguments{
\item{DV}{triplet representation of the half vectorization}

\item{nr}{number of elements in the half vectorization}

\item{V}{Half vectorization}
}
\description{
Half vectorization triplet to matrix triplet
}
