% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.vectorization.R
\name{vectorization}
\alias{JF.PE4}
\alias{findN}
\alias{get.upper.indices}
\alias{nn12}
\alias{vec.to.vech}
\alias{vech.reverse}
\alias{vech.to.vec}
\alias{vectorization}
\title{Vectorization}
\usage{
findN(nv, half.vec = TRUE, check.integer = TRUE)

nn12(n, keep.diag = TRUE)

vech.to.vec(n, keep.diag = TRUE)

get.upper.indices(n, keep.diag = TRUE)

vec.to.vech(n, keep.diag = TRUE)

vech.reverse(V, keep.diag = TRUE, symmetric = TRUE)

JF.PE4(n, p = n, half.vec = FALSE)
}
\arguments{
\item{nv}{length of the vectorization}

\item{half.vec}{Is it a vectorization or a half vectorization}

\item{check.integer}{Should the result be an integer}

\item{n}{the number of rows of the symmetric matrix}

\item{keep.diag}{Is the diagonal kept in the half vectorization}

\item{V}{The vectorization}

\item{symmetric}{Is the matrix symmetric}

\item{M}{Matrix to vectorize}
}
\details{
Function to compute the matrix that will transform M(i, j, k, l) into M(i, l, k, j)
This is commonly known as the commutation matrix
https://en.wikipedia.org/wiki/Commutation_matrix
}
\section{Functions}{
\itemize{
\item \code{findN}: Finds the size of the matrix corresponding to the length of its vectorization

\item \code{nn12}: Computes the length of the half vectorization. n (n + 1) / 2 if we keep the diagonal and n (n - 1) / 2

\item \code{vech.to.vec}: Matrix to go from half vectorization to full vectorization

\item \code{get.upper.indices}: Gets the indices of the upper diagonal

\item \code{vec.to.vech}: Matrix to go from full vectorization to half vectorization

\item \code{vech.reverse}: Takes a half vectorization and builds the corresponding matrix of class \code{Matrix}

\item \code{JF.PE4}: Computes the commuation matrix, i.e matrix K such that K %*%  vec(A) = vec(t(A))
}}

