% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.beautfication.R
\name{wide.screen}
\alias{wide.screen}
\title{Wider Printing}
\usage{
wide.screen(howWide = Sys.getenv("COLUMNS"))
}
\arguments{
\item{howWide}{The width in number of columns to print tables and matrices}
}
\description{
Increases width of object printing
}
\details{
Courtesy of Steve Lianoglou
http://stackoverflow.com/questions/1172485/how-to-increase-the-number-of-columns-using-r-in-linux
}
