% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{with.names}
\alias{with.names}
\alias{eigen.with.names}
\alias{svd.with.names}
\title{More verbose factorization}
\usage{
eigen.with.names(M, ...)

svd.with.names(M, ...)
}
\arguments{
\item{M}{matrix to factorize}
}
\details{
Base methods to factorize matrices, namely \code{eigen} and \code{svd} discard the names of the matrix in their results. These methods add them back.
}
