% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{convert.columns}
\alias{convert.columns}
\title{Converting column types}
\usage{
convert.columns(D, lconv)
}
\arguments{
\item{D}{Table to convert}

\item{lconv}{Mapping between column names and new types}
}
\description{
Converts a set of columns.
}

