% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.table.R
\name{convert.columns}
\alias{convert.columns}
\alias{convert.cols}
\title{Converting column types}
\usage{
convert.columns(D, lconv, ...)

convert.cols(D, col, type___, conv.map = util.type.conv.map, ...)
}
\arguments{
\item{D}{Table to convert}

\item{lconv}{Mapping between column names and new types}

\item{col}{Column to convert}

\item{type___}{Type to convert to}

\item{conv.map}{Map that gives the converting function for each type}
}
\description{
Converting column types
}
\section{Functions}{
\itemize{
\item \code{convert.columns}: Converts a set of columns.

\item \code{convert.cols}: Converts one column to desired type. Using convert.columns is prefered that directly using this function
}}

