% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.tests.R
\name{expect_equal_matrices}
\alias{expect_equal_matrices}
\title{Matrix Equality Expectation}
\usage{
expect_equal_matrices(M1, M2, compare.dimnames = TRUE,
  tol = .Machine$double.eps, agg = TRUE, ...)
}
\arguments{
\item{M1}{Matrix of any form that can be converted to a triplet with function \code{mat.to.triplet}}

\item{M2}{DITTO}

\item{compare.dimnames}{Should dimnames be compared as well}

\item{tol}{Tolerance for equality}

\item{agg}{Should triplet representation that have several entries for the same coordinates be aggregated}

\item{...}{Not used}
}
\description{
Tests whether two matrix represent the same data by converting them to a triplet representation.
}
\details{
Calls to \code{expect_equal} yield to unexpected results. If the goal of the comparison is only to check whether the conceptual matrices represented by M1 and M2 are identical this does the trick.
}

