% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.hash.R
\name{hash.atomic}
\alias{hash.atomic}
\title{Atomic Type Hashing}
\usage{
hash.atomic(x)
}
\arguments{
\item{x}{Vector to be hashed}
}
\description{
Hashes a vector of atomic type (character, numeric, integer, logical, + factor) to an integer
}

