% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.packages.R
\name{install.git}
\alias{install.git}
\title{Install Git}
\usage{
install.git(name, group, url = sprintf("git@www.datasaiyan.com:\%s/\%s.git",
  group, name), install.fun = function(ploc, ...) install.packages(ploc, repos
  = NULL, ...), ...)
}
\arguments{
\item{name}{Name of the package}

\item{group}{Gitlab group to which the package belongs}

\item{url}{Remote url of the package}

\item{install.fun}{Function to use to install the packages}
}
\description{
Installing Jalgos packages with git without using git2r package.
}

