% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsutils-package.R
\docType{package}
\name{jsutils-package}
\alias{jsutils}
\alias{jsutils-package}
\title{jsutils: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Andrew R. Johnson \email{andrew.johnson@arjohnsonau.com} (\href{https://orcid.org/0000-0001-7000-8065}{ORCID})

}
\keyword{internal}
