% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsutils-package.R
\docType{package}
\name{jsutils-package}
\alias{jsutils}
\alias{jsutils-package}
\title{jsutils: 'JavaScript' Utilities for 'R'}
\description{
A collection of popular/useful JavaScript utilities, including the terser minifier, sass compiler, typescript transpiler, and more.
}
\author{
\strong{Maintainer}: Andrew R. Johnson \email{andrew.johnson@arjohnsonau.com} (\href{https://orcid.org/0000-0001-7000-8065}{ORCID})

}
\keyword{internal}
