% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{jsutils-package}
\alias{jsutils}
\alias{jsutils-package}
\title{jsutils: Core util functions}
\description{
A toolbox with functions to perform common tasks.
}
\author{
\strong{Maintainer}: Sébastien Lamy de la Chapelle \email{sebastien@jalgos.com}

Authors:
\itemize{
  \item Julien Murésianu \email{julien@jalgos.com}
}

Other contributors:
\itemize{
  \item Michael Pursey \email{michael@jalgos.com} [contributor]
}

}
