% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.S4.R
\name{mixing-traits}
\alias{mixing-traits}
\alias{add.trait}
\alias{remove.trait}
\title{Adding And Removing Traits}
\usage{
add.trait(object, trait, ...)

remove.trait(object, trait, ...)
}
\arguments{
\item{object}{The object that needs to be improved}

\item{trait}{Name of the trait (an S4 class) that needs to be added or remove from the class definition of the object}

\item{...}{Additional parameters that will be passed on to the initializer of the new.trait}
}
\value{
An object that inherits from its old parent traits as well as the new trait provided
}
\description{
Adding And Removing Traits

Mixing Traits
}
\section{Functions}{
\itemize{
\item \code{add.trait}: An interface to add a new traits to an object. The function will get all the parent traits of the current object and create a new class adding `new.trait` to the list of direct superclasses. It will create a composite object making sure the constructor of the trait is called with `...`

\item \code{remove.trait}: Removes a trait from an object.
}}

