% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.non.finite.R
\docType{methods}
\name{non.finite.elements}
\alias{non.finite.elements}
\alias{non.finite.elements,sparseMatrix-method}
\title{Non finite elements}
\usage{
non.finite.elements(x)

\S4method{non.finite.elements}{sparseMatrix}(x)
}
\description{
Non finite elements
}
