% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{partial.kronecker}
\alias{partial.kronecker}
\title{Partial kronecker product}
\usage{
partial.kronecker(M1, M2, DI, DJ = DI[, list(k = i, l = j, J = I)],
  dims = c(DI[, max(I)], DJ[, max(J)]), half.vec = TRUE, real = TRUE,
  logger = jlogger::JLoggerFactory("Jalgos Algebra"))
}
\description{
Partial kronecker product
}

