% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{partial.kronecker}
\alias{partial.kronecker}
\title{Partial Kronecker Product}
\usage{
partial.kronecker(M1, M2, Il = NULL, Ir = NULL)
}
\arguments{
\item{M1}{a generic matrix that can be converted to a triplet using \code{mat.to.triplet}}

\item{M2}{DITTO}

\item{Il}{The list of row indices in the result that we are interested in keeping. If NULL assumes no projection.}

\item{Ir}{The list of column indices in the result that we are interested in keeping. If NULL assumes no projection.}
}
\description{
Computes a kronecker product for a limited subset of the rows and columns.
}
\details{
Kronecker products produce are very costly in terms of memory usage.
}
