% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\name{partial.kronecker}
\alias{kronecker.proj}
\alias{partial.kronecker}
\title{Partial Kronecker Product}
\usage{
partial.kronecker(M1, M2, Pl = NULL, Pr = NULL)

kronecker.proj(M1, M2, Pl, Pr)
}
\arguments{
\item{M1}{a generic matrix that can be converted to a triplet using \code{mat.to.data.table}}

\item{M2}{DITTO}

\item{Pl}{The list of row indices in the result that we are interested in keeping. If NULL assumes no projection.}

\item{Pr}{The list of column indices in the result that we are interested in keeping. If NULL assumes no projection.}

\item{M1}{a generic matrix that can be converted to a triplet using \code{mat.to.data.table}}

\item{M2}{DITTO}

\item{Pl}{A matrix that will project the rows of M1 %x% M2 into a space of smaller dimension. If not provided will assume identity}

\item{Pr}{A matrix that will project the columns of M1 %x% M2 into a space of smaller dimension. If not provided will assume identity}
}
\details{
Kronecker products produce are very costly in terms of memory usage.
}
\section{Functions}{
\itemize{
\item \code{partial.kronecker}: Computes a kronecker product for a limited subset of the rows and columns.

\item \code{kronecker.proj}: Efficiently projects the result of a kronecker into two (or one) smaller spaces
}}

