% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{positive.negative.value}
\alias{positive.negative.value}
\alias{positive.val}
\alias{negative.val}
\title{Positive and negative value}
\usage{
positive.val(X)

negative.val(X)
}
\arguments{
\item{X}{Numeric vector}
}
\description{
Positive and negative value
}
\section{Functions}{
\itemize{
\item \code{positive.val}: Positive value. Negative values are replaced with 0

\item \code{negative.val}: Negative value. Positive values are replaced with 0
}}

