% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rhistory}
\alias{quit.save}
\alias{rhistory}
\alias{save.rhistory}
\title{Saving Rhistory}
\usage{
save.rhistory(info = "", folder = "Rhistory")

quit.save(...)
}
\arguments{
\item{info}{A name to identify the work you did in the session}

\item{folder}{Folder into which saving the Rhistory files}
}
\details{
Keeping track of your work is very important. These tools allow you to save your entire RHistory so you can comme back to earlier work easily.\cr
The files is saved into the specified directory under the subdirectory current date. The current timestamp is added to the info file so saving your work several time does not overwrite earlier saves. \cr
}
\section{Functions}{
\itemize{
\item \code{save.rhistory}: Saves the entire command history for the current session in a file.

\item \code{quit.save}: Saves the RHistory before exiting
}}

