% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.save.objects.R
\name{save_load}
\alias{save_load}
\alias{save.object}
\alias{read.object}
\title{Saving Loading Objects}
\usage{
save.object(OBJ, file, ..., logger = NULL)

read.object(file, ..., logger = NULL)
}
\arguments{
\item{OBJ}{object to save}

\item{file}{full path of the saving destination}

\item{...}{To be passed to the saving / reading method}

\item{logger}{JLogger}
}
\description{
Saving Loading Objects
}
\section{Functions}{
\itemize{
\item \code{save.object}: Uses `saveRDS` or `qs::qsave` based on the extension of file

\item \code{read.object}: Uses `readRDS` or `qs::qread` based on the extension of file
}}

