% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.data.R
\name{split.write.csv}
\alias{split.write.csv}
\title{Split Write csv}
\usage{
\method{split}{write.csv}(TAB, split.size, export.folder = "output",
  logger = get.util.data.logger())
}
\arguments{
\item{TAB}{Data.table, usually big, to be splitted and exported as csv files}

\item{split.size}{parameter controlling the max size of splits, in number of lines}

\item{export.folder}{parameter for target path to repo to save the csv files}
}
\description{
Allows to export a big data.table as several csv files
}
