% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terser.R
\name{terser}
\alias{terser}
\title{Minify JavaScript code using Terser}
\usage{
terser(code, options = list())
}
\arguments{
\item{code}{A character string containing the JavaScript code to be minified.}

\item{options}{A list of options to pass to Terser for minification.
See the \href{https://terser.org/docs/api-reference/#minify-options}{Terser documentation} for available options.}
}
\value{
A list containing the minified code and any warnings or errors.
}
\description{
This function uses the Terser JavaScript library to minify JavaScript code.
It leverages the QuickJSR package to run JavaScript code within R.
}
\examples{
js_code <- "function add(a, b) { return a + b; }"
terser(js_code, list(sourceMap = TRUE))

}
