% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_matrix.R
\docType{methods}
\name{trim.cov.matrix,ANY-method}
\alias{trim.cov.matrix,ANY-method}
\title{Removing Least Significant Covariance Coefficients}
\usage{
\S4method{trim.cov.matrix}{ANY}(S, Dr = sqDr$diag, D1r = sqDr$inv.diag,
  Dl = sqDl$diag, D1l = sqDl$inv.diag, sqDl = sq.diag.inv(S, tol.inv
  = tol.inv), sqDr = sqDl, ..., mtype = "", tol = 0, tol.inv = 10 *
  .Machine$double.eps, logger = NULL)
}
\arguments{
\item{S}{The covariance matrix}

\item{sqDl}{square root of the diagonal values of the matrix that will be mutltiplied to S from the left side}

\item{sqDr}{square root of the diagonal values of the matrix that will be mutltiplied to S from the right side}

\item{...}{Not used for now}

\item{mtype}{Character vector describing the type of matrix. Used to log information about the process}

\item{tol}{tolerance for the correlation values. Any values under this value will be trimmed out}

\item{tol.inv}{tolerance for inverting a diagonal value. If below this value the entry will be considered null}

\item{logger}{JLogger}

\item{S}{The Matrix}
}
\description{
Will reduce the size of a covariance matrix by removing its least significant coefficients. The trimming will happen in the correlation matrix in order not to overweight the entries with the biggest variance.
}
