% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.vectorization.R
\name{vectorization}
\alias{vectorization}
\alias{findN}
\alias{nn12}
\alias{vech.to.vec}
\alias{get.upper.indices}
\alias{vec.to.vech}
\alias{commutation.matrix}
\alias{block.to.vech}
\title{Vectorization}
\usage{
findN(nv, half.vec = TRUE, check.integer = TRUE)

nn12(n, keep.diag = TRUE)

vech.to.vec(n, keep.diag = TRUE, fmat = default.matrix.builder)

get.upper.indices(
  n,
  i1 = indexation(N = n),
  i2 = indexation(N = n),
  keep.diag = TRUE,
  distributed = mat.vectorization.distributed()
)

vec.to.vech(n, keep.diag = TRUE, fmat = default.matrix.builder)

commutation.matrix(
  n,
  p = n,
  i = indexation(N = n),
  j = indexation(N = p),
  half.vec = FALSE,
  fmat = default.matrix.builder,
  distributed = mat.vectorization.distributed()
)

block.to.vech(
  n1,
  n2 = n1,
  I1 = data.table::CJ(i = 1:n1, j = 1:n1)[i <= j],
  I2 = data.table::CJ(i = 1:n1, j = 1:n2),
  I3 = data.table::CJ(i = 1:n2, j = 1:n2)[i <= j],
  fmat = dsparseMatrix
)
}
\arguments{
\item{nv}{length of the vectorization}

\item{half.vec}{Is it a vectorization or a half vectorization}

\item{check.integer}{Should the result be an integer}

\item{n}{the number of rows of the symmetric matrix}

\item{keep.diag}{Is the diagonal kept in the half vectorization}

\item{M}{Matrix to vectorize}

\item{V}{The vectorization}

\item{symmetric}{Is the matrix symmetric}
}
\description{
Vectorization
}
\details{
Function to compute the matrix that will transform M(i, j, k, l) into M(i, l, k, j)
This is commonly known as the commutation matrix
https://en.wikipedia.org/wiki/Commutation_matrix
}
\section{Functions}{
\itemize{
\item \code{findN}: Finds the size of the matrix corresponding to the length of its vectorization

\item \code{nn12}: Computes the length of the half vectorization. n (n + 1) / 2 if we keep the diagonal and n (n - 1) / 2

\item \code{vech.to.vec}: Matrix to go from half vectorization to full vectorization

\item \code{get.upper.indices}: Gets the indices of the upper diagonal

\item \code{vec.to.vech}: Matrix to go from full vectorization to half vectorization

\item \code{commutation.matrix}: Computes the commuation matrix, i.e matrix K such that K %*%  vec(A) = vec(t(A))

\item \code{block.to.vech}: Computes the matrix that will transform the concatenation of the vectorization (half for diagonal blocks) of each matrix block into the half vectorization of the resulting matrix
}}

